fix(agentbundle): repair Wave 4 CI boundaries
  Move repository-owned acceptance tests out of the sdist suite, correct
  Pages ledger filtering, strengthen adapter diagnostics, and mark Wave 4
  shipped when this PR lands.

  Engine-Change-RFC: RFC-0076
