+++ packages/agentbundle/tests/integration/test_catalogue_wave4_integration.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import json
-import re
 import shutil
 import socket
 import subprocess
@@ -13,7 +12,6 @@ import pytest
 from agentbundle.cli import main
 
 FIXTURE = Path(__file__).parent.parent / "fixtures" / "catalogue_wave4"
-ROOT = Path(__file__).parents[4]
 RESULT_KEYS = {"schema_version", "command", "status", "dry_run", "output", "diagnostics"}
 
 
@@ -334,42 +332,3 @@ def test_command_uses_no_network_or_subprocess(
     monkeypatch.setattr(subprocess, "run", forbidden)
     monkeypatch.setattr(subprocess, "Popen", forbidden)
     assert main(["catalogue", "index", str(root), "--dry-run"]) == 0
-
-
-def _journey_format_section(path: Path) -> str:
-    text = path.read_text(encoding="utf-8")
-    start = text.index("## Journey format")
-    end = text.find("\n## ", start + 3)
-    return text[start:] if end == -1 else text[start:end]
-
-
-def test_journey_format_section_is_portable_and_scaffolded() -> None:
-    source = ROOT / "guides" / "_shared" / "reference" / "catalogue-authoring-standards.md"
-    scaffold = (
-        ROOT
-        / "packages"
-        / "agentbundle"
-        / "agentbundle"
-        / "_data"
-        / "catalogue-scaffold"
-        / "guides"
-        / "_shared"
-        / "reference"
-        / "catalogue-authoring-standards.md"
-    )
-    source_section = _journey_format_section(source)
-    scaffold_section = _journey_format_section(scaffold)
-
-    assert source_section == scaffold_section
-    assert "contracts/catalogue-index.schema.json" in source_section
-    assert "not yet available" not in source_section.lower()
-    for forbidden in (".github/workflows/", "docs/rfc/", "docs/adr/", "docs/specs/"):
-        assert forbidden not in source_section
-    assert re.search(r"(?i)\b(?:RFC|ADR)-\d{4}\b", source_section) is None
-    assert re.search(r"(?i)\bmake\s+\S+", source_section) is None
-
-    for match in re.finditer(r"\[[^]]+\]\(([^)]+)\)", scaffold_section):
-        target = match.group(1).split("#", 1)[0]
-        if not target or target.startswith(("#", "http://", "https://")):
-            continue
-        assert (scaffold.parent / target).resolve().is_file()
